package/pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "dbt_remote_cli"
version = "0.1.0"
description = "cli to run dbt commands on GCP Cloud Run Jobs using dbt-server."
readme = "README.md"
authors = [
  "Emma <a@a.com>"
]
packages = [
    { include = "dbt_remote_cli", from = "src", format = "sdist" }
]
# repository = ""

[tool.poetry.dependencies]
python = "3.11.4"
google-api-core = "2.11.1"
google-cloud-bigquery = "3.11.4"
google-cloud-run = "0.9.0"
google-cloud-core = "2.3.3"
googleapis-common-protos = "1.59.1"
Click = "8.1.3"
requests = "2.31.0"
dbt-bigquery = "1.5.3"
dbt-core = "1.5.2"
pydantic = "1.10.12"

[tool.poetry.scripts]
dbt-remote = 'dbt_remote_cli.cli:cli'
